Add DatabaseSettings tab to SettingsScreen
